Added: new controller for seeding VehicleType & Vehicle seeding now works. Barebone, no additional checks.
