+++ Data/MockupData/SeedVehicles.txt
@@ -4,5 +4,5 @@
     {"UserId": 1, "RegNumber": "BBB123", "Brand": "Polestar", "Model": "Polestar 2", "Color": "Silver Metallic", "NumberOfWheels": 4, "VehicleTypeId": 1},
     {"UserId": 1, "RegNumber": "CCC123", "Brand": "Tesla", "Model": "Model S", "Color": "Red", "NumberOfWheels": 4, "VehicleTypeId": 1},
     {"UserId": 1, "RegNumber": "DDD123", "Brand": "MAN", "Model": "eTGX", "Color": "White", "NumberOfWheels": 10, "VehicleTypeId": 5},
-    {"UserId": 1, "RegNumber": "DDD123", "Brand": "Nimbus", "Model": "T11", "Color": "Mediterranean Blue", "NumberOfWheels": 0, "VehicleTypeId": 3}
+    {"UserId": 1, "RegNumber": "DDD456", "Brand": "Nimbus", "Model": "T11", "Color": "Mediterranean Blue", "NumberOfWheels": 0, "VehicleTypeId": 3}
 ]
