+++ Data/VehicleRepository.cs
@@ -1,4 +1,5 @@
 ﻿using Garage.Models;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
 namespace Garage.Data
@@ -40,8 +41,5 @@ namespace Garage.Data
             _context.Vehicles.Remove(vehicle);
             await _context.SaveChangesAsync();
         }
-
-
-
     }
 }
