+++ Views/Home/Index.cshtml
@@ -3,7 +3,7 @@
 }
 
 <div class="custom-container3 d-flex justify-content-between align-items-start">
-    <form asp-controller="Vehicles" asp-action="AddVehicleSeedsToDb" method="post">
+    <form asp-controller="Seed" asp-action="SeedDatabase" method="post">
         <button type="submit">Seed Vehicles</button>
     </form>
 </div>
